requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap5==2023.10
cycler==0.12.1
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-storages==1.14.2
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
executing==2.0.1
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==20.0.4
h11==0.14.0
idna==3.4
ipython==8.22.1
jedi==0.19.1
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
mypy==1.5.1
mypy-extensions==1.0.0
mysqlclient==2.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
parso==0.8.3
Pillow==10.1.0
ply==3.11
prompt-toolkit==3.0.43
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2020.1
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stone==3.3.1
tomli==2.0.1
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
whitenoise==6.5.0
wsproto==1.2.0
